Attempt to fixup some compiler warnings in Debug mode on GCC and ignore those coming from OpenMC

+++ src/tallies/TallyBase.C
@@ -326,7 +326,7 @@ void
 TallyBase::applyTriggersToLocalTally(openmc::Tally * tally)
 {
   if (_tally_trigger)
-    for (int score = 0; score < _tally_score.size(); ++score)
+    for (int score = 0; score < static_cast<int>(_tally_score.size()); ++score)
       tally->triggers_.push_back({_openmc_problem.triggerMetric((*_tally_trigger)[score]),
                                   _tally_trigger_threshold[score],
                                   false,
